Add cluster-manual-failover-timeout to configure the timeout for manual failover
Allows cluster admins to configure the cluster manual failover timeout as
needed, admins can configure how long a primary would be paused in the
worst case scenario such as a failover timed out due to the insufficient
votes.

The configuration name is cluster-manual-failover-timeout, the unit is
milliseconds, and the default value is 5000.

Signed-off-by: Binbin <binloveplay1314@qq.com>