Fix raxRemove crash at memcpy due to key size exceeds max Rax size
Fix raxRemove crash at memcpy due to key size exceed RAX_NODE_MAX_SIZE.
Note that this could happen when key size was more than 512MB if we
allow it by increasing the default proto-max-bulk-len.
The crash could happen when we recompress the rax after removing a key
due to expiry or DEL while memcpy merge the key that exceed 512MB
limit. While the counting phase has the size check, the actual
compress logic is missing it which lead to this crash.

Signed-off-by: Ram Prasad Voleti <ramvolet@amazon.com>